# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Product Api",
        "description": "A simple CRUD for manage products",
        "contact": {
            "email": "casalaspro.alessandro@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/"
        }
    ],
    "paths": {
        "/api/product": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get all products",
                "operationId": "5cc57bec35d0f2045983cc8dcd00eccd",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Insert new product",
                "operationId": "1d0b0d8724685b7b8f9d87f403bfc4df",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "discount_percentage": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "categories": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "enum": [
                                                "default","technology", "videogames", "photograph", "furniture", "jewels", "edibles", "arts", "clothing", "medicines", "gadgets"
                                            ]
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/product/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Show single product",
                "operationId": "2768d88590e39a7ded1d0032e64dded7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Update single product",
                "description": "Each fild will have a default value. If you don't change that value, we recognize that don't want to make any modify.",
                "operationId": "ec9176d0ba73362b0d0451d9d952f730",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "It has to have 5 chars minimum",
                                        "type": "string",
                                        "default": "default",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "default": "default",
                                        "nullable": true
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary",
                                        "default": "default",
                                        "nullable": true
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "double",
                                        "default": "99999999.99",
                                        "nullable": true
                                    },
                                    "discount_percentage": {
                                        "type": "number",
                                        "format": "double",
                                        "default": "999.99",
                                        "nullable": true
                                    },
                                    "categories": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "default": "default",
                                            "enum": [
                                                "default","technology", "videogames", "photograph", "furniture", "jewels", "edibles", "arts", "clothing", "medicines", "gadgets"
                                            ]
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Product",
            "description": "Product"
        }
    ]
}